clean up screen styles into their own files
