+++ frontend/Screens/Explore.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect } from 'react';
 import {
-  FlatList, Image, Pressable, SafeAreaView, StyleSheet, Text, View,
+  FlatList, Image, Pressable, SafeAreaView, Text, View,
 } from 'react-native';
 import { useSelector } from 'react-redux';
 import { useHistory } from 'react-router-native';
@@ -13,42 +13,7 @@ import CircularThumbnailImage from '../Components/CircularThumbnailImage';
 import {
   randomMountainSelector, unclimbedMountainsSelector, loginSelector, useAppDispatch,
 } from '../store';
-
-const styles = StyleSheet.create({
-  safeArea: {
-    flex: 1,
-    backgroundColor: '#EBF2FA',
-  },
-  title: {
-    flex: 0,
-    marginTop: '5%',
-    marginBottom: '2%',
-    width: '100%',
-    fontSize: 20,
-    fontWeight: 'bold',
-    textAlign: 'center',
-  },
-  pictureTitle: {
-    fontSize: 15,
-    fontWeight: 'bold',
-    textAlign: 'center',
-  },
-  horizontalList: {
-    height: '40%',
-  },
-  unclimbedMountainNames: {
-    color: 'black',
-    textAlign: 'center',
-    width: 100,
-    justifyContent: 'space-around',
-
-  },
-  view: {
-    margin: 10,
-    alignItems: 'flex-end',
-    width: 100,
-  },
-});
+import styles from './styles/exploreStyles';
 
 const Explore = () => {
   const { randomMountainsList } = useSelector(randomMountainSelector);
