+++ frontend/Screens/MountainProfile.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect } from 'react';
 import {
-  SafeAreaView, Text, StyleSheet, Image, View, Pressable,
+  Image, Pressable, SafeAreaView, Text, View,
 } from 'react-native';
 import { useParams, useHistory } from 'react-router-native';
 import { useSelector } from 'react-redux';
@@ -11,33 +11,7 @@ import Header from '../Components/Header';
 import ImageGrid from '../Components/ImageGrid';
 import { globalStyles } from './styles/GlobalStyles';
 import { oneMountainSelector, useAppDispatch } from '../store';
-
-const styles = StyleSheet.create({
-  safeArea: {
-    flex: 1,
-    backgroundColor: '#EBF2FA',
-  },
-  title: {
-    width: '100%',
-    fontSize: 20,
-    fontWeight: 'bold',
-    textAlign: 'center',
-  },
-  imageContainer: {
-    flex: 0,
-  },
-  info: {
-    flex: 0,
-    flexDirection: 'row',
-    justifyContent: 'space-around',
-    alignItems: 'center',
-  },
-  routeButton: {
-    flex: 0,
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-});
+import styles from './styles/mountainProfileStyles';
 
 const MountainProfile = () => {
   const history = useHistory();
