+++ frontend/Screens/styles/loginStyles.tsx
@@ -1,0 +1,18 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: '#EBF2FA',
+  },
+  emailInput: {
+    borderWidth: 1,
+    borderColor: '#427AA1',
+    width: 200,
+    margin: 10,
+    padding: 8,
+    backgroundColor: 'white',
+  },
+});
+
+export default styles;
