+++ frontend/Screens/styles/mapStyles.tsx
@@ -1,0 +1,10 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  safeArea: {
+    flex: 1,
+    backgroundColor: '#EBF2FA',
+  },
+});
+
+export default styles;
